Reset to page 1 when the page count changes
