+++ backend/src/v1/configuration/configuration.controller.ts
@@ -1,6 +1,5 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete, UseGuards } from '@nestjs/common';
+import { Controller, Get, Body, Patch, Param, UseGuards } from '@nestjs/common';
 import { ConfigurationService } from './configuration.service';
-import { CreateConfigurationDto } from './dto/create-configuration.dto';
 import { UpdateConfigurationDto } from './dto/update-configuration.dto';
 import { JwtRoleGuard } from 'src/auth/jwtrole.guard';
 import { ApiTags } from '@nestjs/swagger';
