+++ frontend/src/app/components/mapping/leaflet-map-with-multiple-points.tsx
@@ -38,6 +38,7 @@ const LeafletMapWithMultiplePoints: React.FC<MapProps> = ({ markers }) => {
   return (
     <MapContainer
       style={{ height: "652px", width: "1330px", zIndex: 0 }}
+      className="map-container"
       center={[53.7267, -127.6476]}
       zoom={6}
       ref={mapRef}
